Refactor `replace_wildcard_variables` to use `resolve_variable_references`
